+++ autobot-backend/api/voice_bundle_user.py
@@ -14,12 +14,13 @@ from typing import Optional
 from fastapi import APIRouter, Depends, HTTPException, Request
 from pydantic import BaseModel
 
+from api.voice_bundle_admin import _require_admin
 from api.voice_bundle_constants import (
     BUNDLE_DEFINITIONS,
     VALID_BUNDLES,
     BundleAssignRequest,
 )
-from api.voice_bundle_helpers import _count_tools_for_bundle, _require_admin
+from api.voice_bundle_helpers import _count_tools_for_bundle
 from auth_middleware import get_current_user
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from autobot_shared.logging_manager import get_logger
