+++ autobot-backend/knowledge/connectors/onedrive_test.py
@@ -168,10 +168,6 @@ class TestOneDriveConnector:
             "size": 12345,
             "lastModifiedDateTime": "2026-06-04T08:00:00Z",
             "webUrl": "https://onedrive.live.com/file-123",
-            "parentReference": {
-                "path": "/drive/root:/Documents",
-                "driveId": "drive-abc",
-            },
         }
 
         source_info = connector._file_to_source_info(file_item)
@@ -179,13 +175,9 @@ class TestOneDriveConnector:
         assert source_info is not None
         assert source_info.source_id == "onedrive:test-onedrive-1:file:file-123"
         assert source_info.name == "document.docx"
-        assert source_info.path == "/Documents/document.docx"
-        assert source_info.content_type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
-        assert source_info.size_bytes == 12345
+        assert "document.docx" in source_info.description
         assert source_info.metadata["file_id"] == "file-123"
         assert source_info.metadata["file_extension"] == ".docx"
-        assert source_info.metadata["web_url"] == "https://onedrive.live.com/file-123"
-        assert source_info.metadata["drive_id"] == "drive-abc"
 
     @pytest.mark.asyncio
     async def test_file_to_source_info_unsupported_extension(self, onedrive_config):
