+++ autobot-backend/transcriber/routes/kb.py
@@ -7,7 +7,7 @@
 from fastapi import APIRouter, Depends, HTTPException, Request
 
 from transcriber.database import Database
-from transcriber.deps import DEFAULT_USER, get_db
+from transcriber.deps import get_db
 from transcriber.knowledge.kb_push import push_to_kb
 from transcriber.models import KbPushRequest, KbPushStatus
 from transcriber.routes.export import _build_segment_list
@@ -17,7 +17,7 @@ router = APIRouter(tags=["transcriber-kb"])
 
 def _user_id(request: Request) -> str:
     user = getattr(request.state, "user", None)
-    return user.id if user else DEFAULT_USER
+    return user.id if user else "default"
 
 
 @router.post("/recordings/{recording_id}/kb/push")
